fix(laravel): register error handler without graphql
