+++ src/Laravel/ApiPlatformDeferredProvider.php
@@ -22,6 +22,7 @@ use ApiPlatform\GraphQl\Type\TypeConverterInterface;
 use ApiPlatform\GraphQl\Type\TypesContainerInterface;
 use ApiPlatform\JsonApi\Filter\SparseFieldset;
 use ApiPlatform\JsonApi\Filter\SparseFieldsetParameterProvider;
+use ApiPlatform\Laravel\Controller\ApiPlatformController;
 use ApiPlatform\Laravel\Eloquent\Extension\FilterQueryExtension;
 use ApiPlatform\Laravel\Eloquent\Extension\QueryExtensionInterface;
 use ApiPlatform\Laravel\Eloquent\Filter\BooleanFilter;
@@ -42,11 +43,13 @@ use ApiPlatform\Laravel\Eloquent\State\LinksHandler;
 use ApiPlatform\Laravel\Eloquent\State\LinksHandlerInterface;
 use ApiPlatform\Laravel\Eloquent\State\PersistProcessor;
 use ApiPlatform\Laravel\Eloquent\State\RemoveProcessor;
+use ApiPlatform\Laravel\Exception\ErrorHandler;
 use ApiPlatform\Laravel\Metadata\CacheResourceCollectionMetadataFactory;
 use ApiPlatform\Laravel\Metadata\ParameterValidationResourceMetadataCollectionFactory;
 use ApiPlatform\Laravel\State\ParameterValidatorProvider;
 use ApiPlatform\Laravel\State\SwaggerUiProcessor;
 use ApiPlatform\Laravel\State\ValidateProvider;
+use ApiPlatform\Metadata\IdentifiersExtractorInterface;
 use ApiPlatform\Metadata\InflectorInterface;
 use ApiPlatform\Metadata\Operation\PathSegmentNameGeneratorInterface;
 use ApiPlatform\Metadata\Property\Factory\PropertyMetadataFactoryInterface;
@@ -80,9 +83,11 @@ use ApiPlatform\State\ProcessorInterface;
 use ApiPlatform\State\Provider\ParameterProvider;
 use ApiPlatform\State\Provider\SecurityParameterProvider;
 use ApiPlatform\State\ProviderInterface;
+use Illuminate\Contracts\Debug\ExceptionHandler as ExceptionHandlerInterface;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Support\DeferrableProvider;
 use Illuminate\Support\ServiceProvider;
+use Negotiation\Negotiator;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
@@ -289,6 +294,26 @@ class ApiPlatformDeferredProvider extends ServiceProvider implements DeferrableP
                 $app->make(InflectorInterface::class)
             );
         });
+
+        $this->app->singleton(
+            ExceptionHandlerInterface::class,
+            function (Application $app) {
+                /** @var ConfigRepository */
+                $config = $app['config'];
+
+                return new ErrorHandler(
+                    $app,
+                    $app->make(ResourceMetadataCollectionFactoryInterface::class),
+                    $app->make(ApiPlatformController::class),
+                    $app->make(IdentifiersExtractorInterface::class),
+                    $app->make(ResourceClassResolverInterface::class),
+                    $app->make(Negotiator::class),
+                    $config->get('api-platform.exception_to_status'),
+                    $config->get('app.debug'),
+                    $config->get('api-platform.error_formats')
+                );
+            }
+        );
     }
 
     /**
@@ -329,6 +354,7 @@ class ApiPlatformDeferredProvider extends ServiceProvider implements DeferrableP
             ResourceMetadataCollectionFactoryInterface::class,
             'api_platform.graphql.state_provider.parameter',
             FieldsBuilderEnumInterface::class,
+            ExceptionHandlerInterface::class,
         ];
     }
 }
